UIController Assignment
A modified version of a previous assignment. In the original
assignment, the goal was to create a pathfinding app. This version adds
two other input modes which are directional input and swiping. You can
also take a picture with the camera and use it as a player photo for
the avatar.